Updated question panel
The prompt for the floating panel has been updated with a thermodynamics problem. Indicators have been added in the form of sprites next to the input fields. These will update when players check their answers, and can display either a checkmark (correct), an X (incorrect), or a question mark (improper formatting).