Updated question panel again
Added diagram to question panel showing simple condenser for thermodynamic question.
Diagram has labels for provided question information as well as answers. These are shown at the inputs and outputs of the diagram. Implemented underlying math for the Chiller 2 mission (being used for Demo). Answers have a default tolerance of 1%. Once a correct answer is accepted, the input field is disabled and the answer will show up on the diagram. Once all answers are marked as correct, the GameState will be updated.